Move listing and sale table components
