+++ components/Market/MarketRegion/MarketRegion.tsx
@@ -4,8 +4,8 @@ import { sprintf } from 'sprintf-js';
 import { DataCenter } from '../../../types/game/DataCenter';
 import { Item } from '../../../types/game/Item';
 import { Language } from '../../../types/universalis/lang';
-import ListingsTable from '../../ListingsTable/ListingsTable';
-import SalesTable from '../../SalesTable/SalesTable';
+import ListingsTable from '../ListingsTable/ListingsTable';
+import SalesTable from '../SalesTable/SalesTable';
 import MarketAverages from '../MarketAverages/MarketAverages';
 import MarketCheapest from '../MarketCheapest/MarketCheapest';
 import MarketHistoryGraph from '../MarketHistoryGraph/MarketHistoryGraph';
