+++ components/Market/MarketWorld/MarketWorld.tsx
@@ -4,8 +4,8 @@ import ago from 's-ago';
 import { Item } from '../../../types/game/Item';
 import { World } from '../../../types/game/World';
 import { Language } from '../../../types/universalis/lang';
-import ListingsTable from '../../ListingsTable/ListingsTable';
-import SalesTable from '../../SalesTable/SalesTable';
+import ListingsTable from '../ListingsTable/ListingsTable';
+import SalesTable from '../SalesTable/SalesTable';
 import MarketHistoryGraph from '../MarketHistoryGraph/MarketHistoryGraph';
 import MarketStackSizeHistogram from '../MarketStackSizeHistogram/MarketStackSizeHistogram';
 import NoMarketData from '../NoMarketData/NoMarketData';
